refactor: 코드 리팩토링
- Service 내에 private 메소드 순서 하단으로 이동
- 비즈니스 로직 코드 Service 내로 이동
- 이용내역 문자열 생성 기능 별도 메소드 분리
